[LDN-2018] Update sponsor and images
We've got Google as a sponsor so we need to add that along with
Accenture. I also wanted to move a lot of the images to png rater than a
mixture between jpeg/jpg. Not sure why I did that, I think I was in a
rush.

[LDN-2018] Update speakers images

All speakers images should use same filetype. We also missed a `.jpeg`
type which is meh.